Fixed a bug where the tinder got stuck
This bug was caused by the article receiving an update to the name/title
which led to the href changing. Since the download followed the
redirect, the new key was passed to the frontend and subsequent API
calls were made to the new article, which meant the old article never
got any actions.

Signed-off-by: MaFeLP <60669873+MaFeLP@users.noreply.github.com>